implementing middleware to handle exceptions

+++ GloboTicket.TicketManagement.Application/Features/Events/Commands/DeleteEvent/DeleteEventCommandHandler.cs
@@ -1,5 +1,6 @@
 ﻿using AutoMapper;
 using GloboTicket.TicketManagement.Application.Contracts.Persistance;
+using GloboTicket.TicketManagement.Application.Exceptions;
 using GloboTicket.TicketManagement.Application.Features.Events.Commands.UpdateEvent;
 using GloboTicket.TicketManagement.Domain.Entities;
 using MediatR;
@@ -23,6 +24,9 @@ namespace GloboTicket.TicketManagement.Application.Features.Events.Commands.Dele
         public async Task<Unit> Handle(DeleteEventCommand request, CancellationToken cancellationToken)
         {
             var eventToDelete = await _eventRepository.GetByIdAsync(request.EventId);
+            if (eventToDelete != null) {
+                throw new NotFoundException(nameof(Event), request.EventId);
+            }
             await _eventRepository.DeleteAsync(eventToDelete);
             return Unit.Value;
         }
